Add Chronicler v2 endpoints
